Move Settings properties from MPDevice to a new class (MPSettings)
